Hex edit libnvos.so
Since Android 6.0, dlopen is now POSIX compilant - more specifically, if the
library name contains a slash, it won't interpret it as a relative path from any
LD_LIBRARY_PATH. Instead, it will be interpreted as a path name (relative or absolute)
from the current working directory.
Since our EGL libraries define the library names like 'egl/libGLESv2_tegra'
(more about this later on), and their working directory is usually '/', this
causes a problem.

Luckily, all NVIDIA blobs use the NvOsLibraryLoad function from libnvos.so to load
their libraries. Oddly enough, they usually don't pass the extension (.so) of the
libraries. Instead, there's a check in this function: if the extension isn't present,
it gets appended using an snprintf call - frankly, the format definier (%s.so) is
stored as a string, which we can modify to '/system/lib/%s.so'.
As we know, we can't just add bytes to blobs - we need to overwrite something in
case we want to extend a string. To our luck, there's a string ('NvOsSleepMSInternal()
failure') right after '%s.so', and I was unable to find any references to it - so I
went ahead and overwrote it with our expanded string. Since all NVIDIA libraries are
stored inside /system/lib (with only the EGL blobs being in a subdirectory), this results
in the EGL blobs being able to reach each other, and not affecting others.